requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-environ==0.12.0
django-js-asset==3.1.2
easy-thumbnails==2.10
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
